other_interesting_problems/src/HungarianTest.java
import org.jboss.arquillian.container.test.api.Deployment;
import org.jboss.arquillian.junit.Arquillian;
import org.jboss.shrinkwrap.api.ShrinkWrap;
import org.jboss.shrinkwrap.api.asset.EmptyAsset;
import org.jboss.shrinkwrap.api.spec.JavaArchive;
import org.junit.Assert;
import org.junit.runner.RunWith;

import static org.junit.Assert.*;

@RunWith(Arquillian.class)
public class HungarianTest {
    @Deployment
    public static JavaArchive createDeployment() {
        return ShrinkWrap.create(JavaArchive.class)
                .addClass(Hungarian.class)
                .addAsManifestResource(EmptyAsset.INSTANCE, "beans.xml");
    }

    @org.junit.Test
    public void main() {
        double[][] weight = {{82,83,69,92}, {77,37,49,92}, {11,69,5,86}, {8,9,98,23}};
        Hungarian h = new Hungarian(weight);
        Double res = h.calculate();
        Assert.assertEquals(140, res.doubleValue());


    }

}
